Fix dependency cycle issue
